Complete todolist

+++ app.js
@@ -3,15 +3,10 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const mongoose = require('mongoose');
-//call the date module
-
+const _ = require('lodash');
 
 const app = express();
 const port = 3000;
-
-//array for the item
-// let items = ['Buy Food', 'Cook Food', 'Eat Food'];
-// let workItems = [];
 
 //set the ejs
 app.set('view engine', 'ejs');
@@ -36,6 +31,7 @@ const itemsScheema = {
 
 const Item = mongoose.model('Item', itemsScheema);
 
+//default item
 const item1 = new Item({
     name: "Welcome to  your todolist!"
 });
@@ -50,13 +46,21 @@ const item3 = new Item({
 
 const defaultItems = [item1, item2, item3];
 
+const listSchema = {
+    name: String,
+    items: [itemsScheema]
+};
+
+const List = mongoose.model('List', listSchema);
 
 app.get('/', (req, res) => {
 
     Item.find({}, (err, foundItems) => {
 
+        //if there is no item
         if (foundItems.length === 0) {
 
+            //insert item
             Item.insertMany(defaultItems, (err) => {
                 if (err) {
                     console.log(err);
@@ -80,30 +84,97 @@ app.get('/', (req, res) => {
 
 });
 
+app.get('/:customListName', (req, res) => {
+    //case insensitive
+    const customListName = _.capitalize(req.params.customListName);
+
+    List.findOne({
+        name: customListName
+    }, (err, foundList) => {
+        if (!err) {
+            if (!foundList) {
+                //Create new list
+                const list = new List({
+                    name: customListName,
+                    items: defaultItems
+                });
+
+                list.save();
+
+                res.redirect('/' + customListName);
+            } else {
+                //show an existing list
+                res.render('index', {
+                    listTitle: foundList.name,
+                    items: foundList.items
+                });
+
+            }
+        }
+    });
+
+});
+
 //post function
 app.post('/', (req, res) => {
 
     //grab the value when user type
     const itemName = req.body.newItem;
+    const listName = req.body.list;
 
     const item = new Item({
         name: itemName
     });
 
-    item.save();
+    //if the list in today
+    if (listName === 'Today') {
 
-    res.redirect('/');
+        item.save();
+
+        res.redirect('/');
+    } else {
+        List.findOne({
+            name: listName
+        }, (err, foundList) => {
+            //push it in into foundList
+            foundList.items.push(item);
+            foundList.save();
+            //redirect to the custom list
+            res.redirect('/' + listName);
+        });
+    }
 });
 
+//delete
 app.post('/delete', (req, res) => {
     const checkedItemId = req.body.checkbox;
+    const listName = req.body.listName;
 
-    Item.findByIdAndRemove(checkedItemId, (err) => {
-        if(!err) {
-            console.log('Succesfully deleted checked item.');
-            res.redirect('/');
-        }
-    });
+    //if the list  in today
+    if (listName === 'Today') {
+        Item.findByIdAndRemove(checkedItemId, (err) => {
+            if (!err) {
+                console.log('Succesfully deleted checked item.');
+                res.redirect('/');
+            }
+        });
+    } else {
+        List.findOneAndUpdate({
+            name: listName
+        }, {
+            $pull: {
+                items: {
+                    _id: checkedItemId
+                }
+            }
+        }, (err, foundList) => {
+            if (!err) {
+                res.redirect('/' + listName);
+            }
+        });
+    }
+
+
 });
 
 app.get('/work', (req, res) => {
